# api/openapi/admin/openapi.yaml
openapi: 3.0.3
info:
  title: HVC 后台管理 API
  description: |
    视频转码服务后台管理接口。
    
    包含认证、配置管理、回调管理、RBAC 权限、集群管理、
    转码任务管理、直播频道管理、审计日志等模块。
    所有管理接口需要后台会话认证（Authorization Bearer 或 HttpOnly Cookie）和 RBAC 权限校验。
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 本地开发环境
paths:
...
  /v1/admin/config/publish:
    post:
      summary: 发布配置版本
      tags: [Config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [config_version]
              properties:
                config_version:
                  type: integer
                  format: uint64
                publish_reason:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      config_version:
                        type: integer
                        format: uint64
                      published:
                        type: boolean
                      effective_scope:
                        type: object
                        properties:
                          new_jobs:
                            type: boolean
                          queued_jobs:
                            type: boolean
                          running_jobs:
                            type: boolean
  /v1/admin/config/runtime/server/update:
    post:
      summary: 仅更新 HTTP/WebSocket 对外暴露运行配置
      description: 只接收 enable_http_server 字段，生成新的待发布 runtime config 版本。
      tags: [Config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enable_http_server:
                  type: boolean
                change_summary:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      pending_config_version:
                        type: integer
                        format: uint64
                      published:
                        type: boolean
  /v1/admin/config/runtime/scheduler/update:
    post:
      summary: 仅更新调度策略运行配置
      description: 只接收 scheduler 与请求覆盖相关字段，生成新的待发布 runtime config 版本。
      tags: [Config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                default_profile_id:
                  type: integer
                  format: uint64
                max_global_transcode_sessions:
                  type: integer
                job_lease_ttl_sec:
                  type: integer
                allow_request_override_profile:
                  type: boolean
                allow_request_override_segment_duration:
                  type: boolean
                allow_request_override_hwaccel:
                  type: boolean
                scheduler_loop_interval_ms:
                  type: integer
                require_hardware_encode:
                  type: boolean
                allow_software_decode_fallback:
                  type: boolean
                require_hardware_watermark:
                  type: boolean
                node_cpu_safety_limit_percent:
                  type: integer
                node_memory_safety_limit_percent:
                  type: integer
                node_gpu_safety_limit_percent:
                  type: integer
                change_summary:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      pending_config_version:
                        type: integer
                        format: uint64
                      published:
                        type: boolean
  /v1/admin/config/runtime/worker/update:
    post:
      summary: 仅更新 Worker 运行配置
      description: 只接收 worker 心跳、轮询与单任务上传并发等字段，生成新的待发布 runtime config 版本。
      tags: [Config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                worker_heartbeat_timeout_sec:
                  type: integer
                worker_loop_interval_ms:
                  type: integer
                single_job_upload_concurrency:
                  type: integer
                change_summary:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      pending_config_version:
                        type: integer
                        format: uint64
                      published:
                        type: boolean
  /v1/admin/config/runtime/storage/update:
    post:
      summary: 仅更新存储相关运行配置
      description: 只接收存储域字段，生成新的待发布 runtime config 版本。
      tags: [Config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                storage_type:
                  type: string
                storage_endpoint:
                  type: string
                storage_bucket:
                  type: string
                storage_access_key_id:
                  type: string
                storage_secret_access_key:
                  type: string
                storage_use_ssl:
                  type: boolean
                storage_play_domain:
                  type: string
                storage_flv_domain:
                  type: string
                storage_local_base_path:
                  type: string
                default_storage_id:
                  type: integer
                  format: uint64
                worker_object_prefix:
                  type: string
                change_summary:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      pending_config_version:
                        type: integer
                        format: uint64
                      published:
                        type: boolean
  /v1/admin/config/runtime/callback/update:
    post:
      summary: 仅更新默认回调兜底运行配置
      description: 只接收 enable_callback 和 callback_* 相关字段，生成新的待发布 runtime config 版本。
      tags: [Config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enable_callback:
                  type: boolean
                callback_http_url:
                  type: string
                callback_rpc_endpoint:
                  type: string
                callback_mq_topic:
                  type: string
                change_summary:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      pending_config_version:
                        type: integer
                        format: uint64
                      published:
                        type: boolean
  /v1/admin/config/runtime/mq/update:
    post:
      summary: 仅更新 MQ 消费运行配置
      description: 只接收 MQ 域字段，生成新的待发布 runtime config 版本。
      tags: [Config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enable_mq_consumer:
                  type: boolean
                mq_queue_name:
                  type: string
                mq_host:
                  type: string
                mq_port:
                  type: integer
                mq_username:
                  type: string
                mq_password:
                  type: string
                mq_vhost:
                  type: string
                mq_consumer_tag:
                  type: string
                mq_prefetch_count:
                  type: integer
                change_summary:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      pending_config_version:
                        type: integer
                        format: uint64
                      published:
                        type: boolean
  /v1/admin/config/runtime/grpc/update:
    post:
      summary: 仅更新对外 public gRPC 运行配置
      description: 只接收 public gRPC 域字段，生成新的待发布 runtime config 版本。
      tags: [Config]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enable_grpc_server:
                  type: boolean
                grpc_listen_address:
                  type: string
                public_grpc_registry_id:
                  type: integer
                  format: uint64
                change_summary:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      pending_config_version:
                        type: integer
                        format: uint64
                      published:
                        type: boolean
...
  /v1/admin/config-center/enabled:
    post:
      summary: 启用/禁用 bootstrap 配置源绑定
      description: 启用或禁用外部 bootstrap 配置源元数据，不影响已发布 runtime config 的热更新链路；响应会显式标明 config_scope=bootstrap、affects_runtime=false。
      tags: [ConfigCenter]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [binding_id, enabled]
              properties:
                binding_id:
                  type: integer
                  format: uint64
                enabled:
                  type: boolean
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      binding_id:
                        type: integer
                        format: uint64
                      enabled:
                        type: boolean
                      config_scope:
                        type: string
                      affects_runtime:
                        type: boolean
...